pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taman-kehati-backend"
version = "0.1.0"
description = "FastAPI backend for Taman Kehati"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.32.0",
  "sqlalchemy==2.0.35",
  "pydantic==2.9.2",
  "pydantic-settings==2.6.1",
  "psycopg[binary]==3.2.1",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "python-multipart==0.0.12",
  "alembic==1.13.3",
  "geoalchemy2==0.15.2",
  "shapely==2.0.6",
  "python-dotenv==1.0.1",
]



